Update GUI printg to work.
