+++ src/main/java/compal/logic/LogicManager.java
@@ -7,6 +7,7 @@ import compal.logic.command.exceptions.CommandException;
 import compal.logic.parser.ParserManager;
 import compal.logic.parser.exceptions.ParseException;
 import compal.model.tasks.Task;
+import compal.ui.UiUtil;
 
 import java.util.ArrayList;
 
@@ -16,6 +17,7 @@ import java.util.ArrayList;
 public class LogicManager {
     public static final String MESSAGE_ERROR_COMMAND = "An error encountered while executing command.";
     public static final String BYE_TOKEN = "bye.";
+    private final UiUtil uiUtil;
 
     private Messages messages;
     private ParserManager parserManager;
@@ -23,6 +25,7 @@ public class LogicManager {
     public LogicManager() {
         this.parserManager = new ParserManager();
         this.messages = new Messages();
+        this.uiUtil = new UiUtil();
     }
 
     /**
@@ -33,7 +36,7 @@ public class LogicManager {
 
         Command command = parserManager.processCmd(fullCommand);
         CommandResult cmdResult = command.commandExecute(taskArrList);
-        System.out.println(cmdResult.feedbackToUser);
+        uiUtil.printg(cmdResult.feedbackToUser);
         if (cmdResult.feedbackToUser.equals(BYE_TOKEN)) {
             System.exit(0);
         }
